requirements.txt
Django==1.11.2
django-mptt==0.8.7
django-ratelimit==1.0.1
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
pytz==2017.2
